requirements.txt
opencv-python
numpy
tqdm
Pillow